Declare test env vars in the podspec to make them available for tests
To pass env vars with xcodebuild to the tests we have to declare them
in the test scheme and set their values to be read from the build
settings.

To pass the credentials in Xcode to the tests the env vars can be
update in the test scheme. Make sure not to commit the credentials!